Added a Quetzal example
Quetzal is a language that I am developing, and in an attempt to improve the design, I decided to write the raymarching demo in Quetzal. It cannot be compiled yet as the language design is the deciding factor for how elaborate the parser has to be, and since I'm palling on using my own parser generator for Quetzal, I will have to upgrade my parser generator to be able to parse the language before beginning work on the Compiler.
As or Quetzal itself, I'm planning to make it something close to C/C++ with 4 main goals:
- No keywords! They're just identifiers stolen from the user.
- More built-in stuff so you have to rely less on libraries.
- More convenient than C/C++. This includes stuff like: Support for functions returning multiple values, python-style list comprehensions (they can be trivially vectorized), Tuple unpacking, etc.
- Clear up some ideological clutter found in C/C++